+++ lib/composer/installed.php
@@ -7,9 +7,18 @@
         'type' => 'prestashop-module',
         'install_path' => __DIR__ . '/../../',
         'aliases' => array(),
-        'dev' => true,
+        'dev' => false,
     ),
     'versions' => array(
+        'ddtraceweb/monolog-parser' => array(
+            'pretty_version' => '1.3.1',
+            'version' => '1.3.1.0',
+            'reference' => 'd6715cb28700d691b20d713626856aacd3d5606c',
+            'type' => 'library',
+            'install_path' => __DIR__ . '/../ddtraceweb/monolog-parser',
+            'aliases' => array(),
+            'dev_requirement' => false,
+        ),
         'monext/monext-php' => array(
             'pretty_version' => '4.77.0',
             'version' => '4.77.0.0',