+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ===
 <h2>released on 2024-11-07</h2>
 <ul>
+<li>Update SDK dependency to fit rebranding monext/monext-php</li>
+<li>Simplify secondary contracts configuration</li>
+<li>Update logos</li>
 </ul>
 
 2.3.5